third_party/WebKit/LayoutTests/paint/deprecation/css-clip-and-clip-path-expected.txt

PASS internals.isUseCounted(document, ClipPathOfPositionedElement) is true
PASS internals.isUseCounted(document, ClipCssOfPositionedElement) is true
PASS successfullyParsed is true

TEST COMPLETE
